requirements.txt
nose
sphinx
xbee
paho-mqtt